Androidでのthreadの実装
